REMOVE : component 폴더 하나로 관리할수 있도록 정리
