+++ app/components/custom/basic_info_dropdown/CheckBoxGrayOut.tsx
