+++ app/components/custom/common/DropDownBig.tsx
@@ -3,7 +3,7 @@ import {
   AccordionContent,
   AccordionItem,
   AccordionTrigger,
-} from '@/components/ui/accordion';
+} from '@/app/components/ui/accordion';
 import Switch from './DropDownBIgSwitch';
 import { DropDownBigProps } from '@/type/components';
