+++ app/components/custom/wedding_date_dropdown/DropDownTime.tsx
@@ -4,7 +4,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
+} from '@/app/components/ui/select';
 
 interface DropDownTimeProps {
   selectedValue: string;
