+++ app/page.tsx
@@ -1,5 +1,11 @@
 'use client';
 
+import Wedding_Date_Dropdown from '@/app/components/custom/wedding_date_dropdown/Wedding_Date_Dropdown';
+
 export default function Home() {
-  return <>page</>;
+  return (
+    <>
+      <Wedding_Date_Dropdown />
+    </>
+  );
 }
